added symlink support based on the work done by @urbaniak
